+++ fflauncher/src/main/java/tv/hd3g/fflauncher/about/FFAboutDevice.java
@@ -23,10 +23,11 @@ public class FFAboutDevice extends FFAboutFormat {
 	static List<FFAboutDevice> parseDevices(final List<String> lines) {
 		return lines.stream()
 				.map(String::trim)
+				.dropWhile(line -> line.startsWith("-") == false && line.endsWith("-") == false)
+				.filter(line -> line.startsWith("-") == false && line.endsWith("-") == false)
 				.filter(line -> (line.toLowerCase().startsWith("Devices:".toLowerCase()) == false))
 				.filter(line -> (line.toLowerCase().startsWith("D. = Demuxing supported".toLowerCase()) == false))
 				.filter(line -> (line.toLowerCase().startsWith(".E = Muxing supported".toLowerCase()) == false))
-				.filter(line -> (line.startsWith("--") == false))
 				.map(FFAboutDevice::new)
 				.toList();
 	}