+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # All Medialib projects upgrades needs
 
+Correct bugs with FFmpeg v7:
+
+ - Parse EBUR128 from lavfi metadata instead of strerr
+ - Not parse anything from stderr
+ - Set ffmpeg loglevel to warn
+ - Correct ffprobe result SAX entries (remove `coded_picture_number` and `display_picture_number`, rename `pkt_duration` to `duration`, and `pkt_duration_time` to `duration_time`, but keep previous behavior compatible)
+ - Correct FFAbout with the new ffmpeg output syntax
+
 ## 15.0.2
 
 Update FFprobeJAXB to handle to ffprobe XML v7.0 #88